pyproject.toml
[tool.poetry]
name = "mnist-etc"
version = "0.1.0"
description = "Project for MLOps course"
authors = ["Petros Osetros <petr@osetr.lake>"]
license = "BSD3"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
tensorflow = "<2.14.0"
tqdm = "^4.66.1"
pillow = "^10.0.1"
tensorflow-privacy = "<0.8.11"
dvc = "^3.30.1"
dvc-gdrive = "^2.20.0"
hydra-core = "^1.3.2"
mlflow = "^2.8.1"
onnx = "<1.15.0"
tf2onnx = "^1.15.1"
mlserver = "^1.3.5"
onnxruntime = "^1.16.3"
mlserver-mlflow = "^1.3.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[tool.black]
line-length = 90

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
